# backend/docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  ? github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_TelemetryRecordResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.TelemetryRecordResponse'
        type: array
      total_items:
        type: integer
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountResponse:
    properties:
      count:
        example: 12
        type: integer
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse:
    properties:
      details:
        example: node_class_id is required
        type: string
      error:
        example: Invalid Format
        type: string
      message:
        example: Please select a valid node class.
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.TelemetryRecordResponse:
    properties:
      created_at:
        example: "2026-07-28T08:15:00Z"
        type: string
      id:
        example: 918273
        type: integer
      metric_name:
        example: water_temperature_celsius
        type: string
      node_device_id:
        example: ESP32-BARISTA-07
        type: string
      payload:
        type: object
      payload_schema_name:
        example: brew_command
        type: string
      payload_schema_version:
        example: 1
        type: integer
      recorded_at:
        example: "2026-07-28T08:14:58Z"
        type: string
    type: object
...
info:
  contact: {}
  description: Nusapala Things is the public HTTP and MQTT entrypoint for IoT telemetry,
    action dispatch, and firmware OTA.
  title: Nusapala Things
  version: "1.0"
paths:
...
  /v1/telemetry-records:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Telemetry Record Delete
      tags:
      - Telemetry
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_TelemetryRecordResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Telemetry Record List
      tags:
      - Telemetry
securityDefinitions:
  BearerAuth:
    description: Paste the access token. The API accepts "Bearer " followed by the
      access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"